requirements/base.txt
Django==1.6.5
South==0.8.4
django-classy-tags==0.5.1
django-extensions==1.3.8
django-sekizai==0.7
django-timezone-field==1.0
google-api-python-client==1.2
httplib2==0.9
oauth2client==1.2
oauthlib==0.6.1
psycopg2==2.5.3
python-openid==2.2.5
pytz==2014.4
requests==2.3.0
requests-oauthlib==0.4.0
six==1.7.2
wsgiref==0.1.2
